Fixed code for future build

+++ Procedural-Midi/MidiAsset/Source/Midi/Classes/Util/MidiProcessor.h
@@ -46,7 +46,7 @@ protected:
 
 private:
 	TArray<TArray<MidiEvent*>::TIterator> mCurrEvents;
-	long long mLastMs;
+	uint32 mLastMs;
 	int mCurrentTrack;
 	MidiEventListener* mListener;
 };
